% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtools-package.R
\docType{package}
\name{buildtools-package}
\alias{buildtools}
\alias{buildtools-package}
\title{buildtools: buildtools R Package}
\description{
Tools for isolating, detecting dependencies,
    scaffolding, and building deployable, reproducible developer
    environments in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/buildtools}
  \item Report bugs at \url{https://github.com/jimbrig/buildtools/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimbrig2011@outlook.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
