% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{check_docker_running}
\alias{check_docker_running}
\title{Check that the Docker Daemon/Service is Running}
\usage{
check_docker_running()
}
\value{
Logical TRUE/FALSE
}
\description{
Check that the Docker Daemon/Service is Running
}
