% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{Create Dockerfile from}
\usage{
create_dockerfile(
  path = getwd(),
  base_image = "rocker/r-ver:latest",
  app_config = "default",
  mainainer = whoami::whoami()["fullname"],
  date = Sys.Date(),
  additional_r_commands = NULL,
  port = "8080"
)
}
\arguments{
\item{path}{defaults to getwd()}

\item{base_image}{base image for \code{FROM} in Dockerfile}

\item{app_config}{app config, defaults to \code{default}}

\item{date}{date - defaults to \code{\link[base:Sys.time]{base::Sys.Date()}}}

\item{additional_r_commands}{any additional r commands to include in image}

\item{maintainer}{maintainer - defaults to \code{fullname} from \code{\link[whoami:whoami]{whoami::whoami()}}}

\item{packages}{R package dependencies returned from \code{\link[=get_package_deps]{get_package_deps()}}}

\item{sysreqs}{logical - should sysreqs be included in image (via \code{\link[=get_sysreqs]{get_sysreqs()}})}
}
\value{
invisibly returns list of package and system dependencies
}
\description{
Create Dockerfile from
}
