% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker.R
\name{create_dockerignore}
\alias{create_dockerignore}
\title{Create .dockerignore from template}
\usage{
create_dockerignore(open = TRUE)
}
\arguments{
\item{open}{logical - open file?}
}
\description{
Create .dockerignore from template
}
