% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdeps.R
\name{create_pkgdeps_file}
\alias{create_pkgdeps_file}
\title{Create a \code{deps.yml} Package Dependency File}
\usage{
create_pkgdeps_file(deps = NULL, app_dir = "shiny_app", ...)
}
\arguments{
\item{deps}{list returned from \code{\link[=get_pkgdeps]{get_pkgdeps()}}; leave null to run \code{\link[=get_pkgdeps]{get_pkgdeps()}}.}

\item{app_dir}{the application's path}

\item{...}{
  Arguments passed on to \code{\link[=get_pkgdeps]{get_pkgdeps}}
  \describe{
    \item{\code{ignores}}{packages to ignore (i.e. remotes)}
    \item{\code{verbose}}{logical}
  }}
}
\value{
unnamed list of packages
}
\description{
This function creates a specifically structured \code{YAML} file to house a Shiny
App's dependencies.
}
