% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdeps.R
\name{get_package_deps}
\alias{get_package_deps}
\title{Get R Package Dependencies}
\usage{
get_package_deps(
  path = getwd(),
  write_yaml = TRUE,
  write_r = TRUE,
  include_versions = TRUE
)
}
\arguments{
\item{path}{path to directory}

\item{write_yaml}{logical - should \code{deps.yaml} be created?}

\item{write_r}{logical - should \code{deps.R} be created?}

\item{include_versions}{logical - should package versions and github referenced be included?}
}
\value{
silently returns a data.frame with R package details
}
\description{
This function takes a path to a directory and parses the code from all
\code{.R} and \code{.Rmd} files, retrieving any detected package dependencies, and
optionally outputs a \code{deps.yaml} and \code{deps.R} file.
}
