% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdeps.R
\name{get_pkgdeps}
\alias{get_pkgdeps}
\title{Get Package Dependencies}
\usage{
get_pkgdeps(app_dir = "shiny_app", ignores = NULL, verbose = TRUE)
}
\arguments{
\item{app_dir}{directory to scan}

\item{ignores}{packages to ignore (i.e. remotes)}

\item{verbose}{logical}
}
\value{
list of package dependencies
}
\description{
Get Package Dependencies
}
