% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{get_sysreqs}
\alias{get_sysreqs}
\title{Get System Requirements for R Packages}
\usage{
get_sysreqs(pkgs)
}
\arguments{
\item{pkgs}{R package dependencies as a character vector}
}
\value{
character vector
}
\description{
Get System Requirements for R Packages
}
\examples{
\dontrun{
get_sysreqs("tidyverse")
get_sysreqs(c("plumber", "rmarkdown"))
}
}
