% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildtools.R
\name{find_logo}
\alias{find_logo}
\alias{vignettes_base64}
\alias{commit_info_base64}
\alias{sysdeps_base64}
\alias{install_sysdeps}
\alias{install_dependencies}
\alias{read_description_field}
\alias{get_ostype}
\alias{get_maintainer_info}
\alias{get_gitstats}
\alias{get_gitstats_base64}
\alias{render_readme}
\alias{generate_cff}
\alias{maintainer_info_base64}
\alias{list_assets}
\title{Build tools}
\usage{
find_logo(path, git_url, subdir = "")

vignettes_base64(repo, pkg = basename(repo), subdir = "")

commit_info_base64(repo = ".")

sysdeps_base64(pkg)

install_sysdeps(path = ".")

install_dependencies(path = ".")

read_description_field(field, path = ".")

get_ostype(path = ".")

get_maintainer_info(path = ".")

get_gitstats(repo, pkgdir, url)

get_gitstats_base64(repo, pkgdir, url)

render_readme(repo, pkgdir, outdir)

generate_cff(path, outdir)

maintainer_info_base64(path = ".")

list_assets(path)
}
\arguments{
\item{path}{root directory of package}

\item{git_url}{of the git repository}

\item{subdir}{path within the git repo where the pkg is}

\item{repo}{path to the git repository}

\item{pkg}{name of the installed package}

\item{field}{which field from the description to show}
}
\description{
Get some extra info about packages.
}
