% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysreqs.R
\name{get_sysreqs}
\alias{get_sysreqs}
\alias{get_sysreqs.default}
\alias{get_sysreqs.dep}
\alias{get_sysreqs.rhub}
\alias{get_sysreqs.pak}
\title{Get sysreqs generic}
\usage{
get_sysreqs(pkgs, platform = NULL, method = NULL, ...)

\method{get_sysreqs}{default}(pkgs, ...)

\method{get_sysreqs}{dep}(pkgs, ...)

\method{get_sysreqs}{rhub}(pkgs = NULL, platform = "linux-x86_64-debian-gcc")

\method{get_sysreqs}{pak}(pkgs, ...)
}
\arguments{
\item{pkgs}{Character vector of R packages to check for.}

\item{platform}{Defaults to "linux-x86_64-debian-gcc" for Docker images.}

\item{method}{Method to use. Should be one of pak, dep, rhub, rstudio, sysreqs.}

\item{...}{For methods}
}
\value{
Character vector of system requirements associated with the supplied packages.
}
\description{
Retrieve R Package system requirements using a suite of various methods.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default Method: This method uses the \code{sysreqs} package
directly by calling the function \code{\link[sysreqs:sysreqs]{sysreqs::sysreqs()}} and using package
\code{DESCRIPTION} files.

\item \code{dep}: \code{dep} Method: This method uses \code{\link[dep:get_sysreqs]{dep::get_sysreqs()}}
to retrieve system dependencies.

\item \code{rhub}: R-Hub API Method: This method of the \code{get_sysreqs}
generic performs a \code{GET} request to the R-Hub SysReqs API at path:
\verb{https://sysreqs.r-hub.io/pkg/<pkg1>,<pkg2>,<pkg3>/<platform>} using
\code{\link[httr:GET]{httr::GET()}}.

\item \code{pak}: \code{pak} Method: This method wraps \code{\link[pak:local_system_requirements]{pak::pkg_system_requirements()}}
to retrieve R package system requirements.
}}

\note{
Any packages supplied should be installed on local system to allow this
function to work properly.
}
\examples{
pkgs <- c("magick", "tinytex")
get_sysreqs(pkgs, method = "rhub")
}
\seealso{
\code{\link[dep:get_sysreqs]{dep::get_sysreqs()}}, \code{\link[pak:local_system_requirements]{pak::pkg_system_requirements()}}
}
\keyword{sysreqs}
